WSAD_Models/AnomalyCLIP/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1
torchvision==0.14.1
pytorch-lightning==1.8.3
torchmetrics==0.11.0
hydra-core==1.3.1
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
wandb
pyrootutils
pre-commit
rich
pytest
pillow==9.0.1
opencv-python==4.5.5.64
axial_attention
seaborn
dotmap
ftfy
einops
regex
tqdm